Port to Clean Architecture
Initial attempt at migrating to Clean Architecture, might require further changes
