+++ lib/core/singletons/package_info_singleton.dart
