+++ lib/core/singletons/shared_preferences_singleton.dart
