+++ lib/features/playlist/data/repositories/playlist_repostiory_impl.dart
@@ -1,0 +1,11 @@
+import 'package:vidya_music/features/playlist/domain/entities/playlist.dart';
+import 'package:vidya_music/features/playlist/domain/repositories/playlist_repository.dart';
+
+class PlaylistRepositoryImpl extends PlaylistRepository {
+  PlaylistRepositoryImpl();
+
+  @override
+  Future<Playlist?> getPlaylist() async {
+    return null;
+  }
+}
