+++ lib/features/playlist/data/repositories/track_repository_impl.dart
@@ -1,0 +1,11 @@
+import 'package:vidya_music/features/playlist/domain/entities/track.dart';
+import 'package:vidya_music/features/playlist/domain/repositories/track_repository.dart';
+
+class TrackRepositoryImpl extends TrackRepository {
+  TrackRepositoryImpl();
+
+  @override
+  Future<List<Track>?> getTracks() async {
+    return null;
+  }
+}
