+++ lib/features/playlist/domain/entities/playlist.dart
