+++ lib/features/playlist/domain/repositories/config_repository.dart
@@ -1,0 +1,7 @@
+// ignore_for_file: one_member_abstracts
+
+import 'package:vidya_music/features/playlist/domain/entities/config.dart';
+
+abstract class ConfigRepository {
+  Future<Config?> getConfig();
+}
